# TryAddWithoutValidation for multiple values could be more efficient