--- a/src/libraries/System.Net.Http/src/System/Net/Http/Headers/HttpHeaders.cs
+++ b/src/libraries/System.Net.Http/src/System/Net/Http/Headers/HttpHeaders.cs
@@ -168,108 +168,58 @@ namespace System.Net.Http.Headers
         internal bool TryAddWithoutValidation(HeaderDescriptor descriptor, IEnumerable<string?> values)
         {
             ArgumentNullException.ThrowIfNull(values);
-            string?[]? valuesArray = values as string?[];
-            IList<string?>? valuesList = values as IList<string?>;
 
-            // the count is null when values is not a array nor IList
-            int? count = valuesArray != null ? valuesArray.Length : valuesList?.Count;
-
-            if (count == 0)
+            if (values is IList<string?> valuesList && valuesList.Count > 0)
             {
-                return true;
-            }
+                int count = valuesList.Count;
 
-            // read the store of header values
-            // The header store contain a single raw string value when header values are single.
-            // The header store contain HeaderStoreItemInfo that wraps around a List<string> with header values when they are more than once
+                // reads the store of header values
+                // The header store contain a single raw string value when header values are single.
+                // The header store contain HeaderStoreItemInfo that wraps around a List<string> with header values when they are more than once
 
-            ref object? storeValueRef = ref GetValueRefOrAddDefault(descriptor);
-            object? storeValue = storeValueRef;
+                ref object? storeValueRef = ref GetValueRefOrAddDefault(descriptor);
+                object? storeValue = storeValueRef;
 
-            // check whether the header store contains already a value
-            // or header values count is more than one
-            // => allocate a HeaderStoreItemInfo on store that wraps around a List<string> of header values
-            if (storeValue is not null || (count.HasValue && count.Value > 1))
-            {
-                if (storeValue is not HeaderStoreItemInfo info)
+                // check whether the header store contains already a value
+                // or header values count is more than one
+                // => allocate a HeaderStoreItemInfo on store that wraps around a List<string> of header values
+                if (storeValue is not null || count > 1)
                 {
-                    storeValueRef = info = new HeaderStoreItemInfo { RawValue = storeValue };
-                }
-
-                object? rawValue = info.RawValue;
-                if (rawValue is not List<string> rawValues)
-                {
-                    info.RawValue = rawValues = new List<string>();
-
-                    if (rawValue != null)
+                    if (storeValue is not HeaderStoreItemInfo info)
                     {
-                        if (count.HasValue)
+                        storeValueRef = info = new HeaderStoreItemInfo { RawValue = storeValue };
+                    }
+
+                    object? rawValue = info.RawValue;
+                    if (rawValue is not List<string> rawValues)
+                    {
+                        info.RawValue = rawValues = new List<string>();
+
+                        if (rawValue != null)
                         {
-                            rawValues.EnsureCapacity(count.Value + 1);
+                            rawValues.EnsureCapacity(count + 1);
+                            rawValues.Add((string)rawValue);
                         }
-
-                        rawValues.Add((string)rawValue);
                     }
-                }
 
-                if (count.HasValue)
-                {
-                    rawValues.EnsureCapacity(count.Value);
-                }
+                    rawValues.EnsureCapacity(count);
 
-                if (valuesArray != null)
-                {
-                    for (int i = 0; i < valuesArray.Length; i++)
-                    {
-                        rawValues.Add(valuesArray[i] ?? string.Empty);
-                    }
-                }
-                else if (valuesList != null)
-                {
-                    for (int i = 0; i < valuesList!.Count; i++)
+                    for (int i = 0; i < count; i++)
                     {
                         rawValues.Add(valuesList[i] ?? string.Empty);
                     }
                 }
                 else
                 {
-                    foreach (string? value in values)
-                    {
-                        rawValues.Add(value ?? string.Empty);
-                    }
+                    // when header values are single the store contains just the header value
+                    storeValueRef = valuesList[0];
                 }
-            }
-            else if (count == 1)
-            {
-                // when header values are single the store contains just the header value
-                storeValueRef = valuesArray != null ? valuesArray[0] : (valuesList != null ? valuesList![0] : null);
             }
             else
             {
-                // handles the case when header values count is unknown because the values are abstracted by IEnumerable<string>
                 foreach (string? value in values)
                 {
-                    if (storeValueRef is null)
-                    {
-                        storeValueRef = value;
-                    }
-                    else
-                    {
-                        if (storeValueRef is not HeaderStoreItemInfo info)
-                        {
-                            storeValueRef = info = new HeaderStoreItemInfo { RawValue = storeValueRef };
-                        }
-
-                        object? rawValue = info.RawValue;
-
-                        if (rawValue is not List<string> rawValues)
-                        {
-                            info.RawValue = rawValues = new List<string>(2);
-                            rawValues.Add(rawValue == null ? string.Empty : (string)rawValue);
-                        }
-
-                        rawValues.Add(value ?? string.Empty);
-                    }
+                    TryAddWithoutValidation(descriptor, value ?? string.Empty);
                 }
             }