--- a/src/libraries/System.Net.Http/tests/UnitTests/Headers/HttpHeadersTest.cs
+++ b/src/libraries/System.Net.Http/tests/UnitTests/Headers/HttpHeadersTest.cs
@@ -2575,36 +2575,6 @@ namespace System.Net.Http.Tests
         }
 
         [Fact]
-        public void TryAddWithoutValidation_OneValidValueHeader_UseSpecialArrayImplementation()
-        {
-            const string Name = "customHeader1";
-            const string Value = "Value1";
-
-            var response = new HttpResponseMessage();
-            Assert.True(response.Headers.TryAddWithoutValidation(Name, new [] { Value }));
-
-            Assert.True(response.Headers.Contains(Name));
-
-            Assert.True(response.Headers.TryGetValues(Name, out IEnumerable<string> values));
-            Assert.Equal(Value, values.Single());
-        }
-
-        [Fact]
-        public void TryAddWithoutValidation_ThreeValidValueHeader_UseSpecialArrayImplementation()
-        {
-            const string Name = "customHeader1";
-            string[] expectedValues = ["Value1", "Value2", "Value3"];
-
-            var response = new HttpResponseMessage();
-            Assert.True(response.Headers.TryAddWithoutValidation(Name, expectedValues));
-
-            Assert.True(response.Headers.Contains(Name));
-
-            Assert.True(response.Headers.TryGetValues(Name, out IEnumerable<string> values));
-            Assert.True(expectedValues.SequenceEqual(values));
-        }
-
-        [Fact]
         public void TryAddWithoutValidation_OneValidValueHeader_UseGenericImplementation()
         {
             const string Name = "customHeader1";