Fixes #64049
perfomance test dotnet/performance#4240

```
BenchmarkDotNet v0.13.13-nightly.20240311.145, Windows 11 (10.0.22631.3593/23H2/2023Update/SunValley3)
12th Gen Intel Core i7-12700H, 1 CPU, 20 logical and 14 physical cores
.NET SDK 9.0.100-preview.4.24267.66
  [Host]     : .NET 9.0.0 (9.0.24.26619), X64 RyuJIT AVX2
  Job-RIQMYG : .NET 9.0.0 (42.42.42.42424), X64 RyuJIT AVX2
  Job-UYKONU : .NET 9.0.0 (42.42.42.42424), X64 RyuJIT AVX2

PowerPlanMode=00000000-0000-0000-0000-000000000000  IterationTime=250ms  MaxIterationCount=20  
MinIterationCount=15  WarmupCount=1  
```

| Method     | Toolchain | data    | Mean     | Error    | Ratio | Allocated | Alloc Ratio |
|----------- |---------- |-------- |---------:|---------:|------:|----------:|------------:|
| AddHeaders | pr        | Array   | 45.91 ns | 1.350 ns |  0.54 |     336 B |        0.69 |
| AddHeaders | main      | Array   | 84.59 ns | 1.230 ns |  1.00 |     488 B |        1.00 |
|            |           |         |          |          |       |           |             |
| AddHeaders | pr        | Hashset | 86.72 ns | 1.129 ns |  0.97 |     496 B |        1.00 |
| AddHeaders | main      | Hashset | 89.18 ns | 1.356 ns |  1.00 |     496 B |        1.00 |
|            |           |         |          |          |       |           |             |
| AddHeaders | pr        | List    | 45.95 ns | 0.730 ns |  0.54 |     336 B |        0.68 |
| AddHeaders | main      | List    | 85.28 ns | 0.809 ns |  1.00 |     496 B |        1.00 |