@pedrobsaila 2024-05-31
New results after adjusting performance test and TryAddWithoutValidation method :

| Method                       | Toolchain                                                                                                      | data                | Mean     | Error    | StdDev   | Median   | Min      | Max      | Ratio | Gen0   | Allocated | Alloc Ratio |
|----------------------------- |--------------------------------------------------------------------------------------------------------------- |-------------------- |---------:|---------:|---------:|---------:|---------:|---------:|------:|-------:|----------:|------------:|
| **TryAddWithoutValidationTests** | **PR** | **Array** | **34.59 ns** | **0.360 ns** | **0.300 ns** | **34.54 ns** | **34.13 ns** | **35.23 ns** |  **0.47** | **0.0102** |     **128 B** |        **0.46** |
| TryAddWithoutValidationTests | master      | Array | 73.25 ns | 0.744 ns | 0.696 ns | 73.29 ns | 72.11 ns | 74.30 ns |  1.00 | 0.0222 |     280 B |        1.00 |
|                              |                                                                                                                |                     |          |          |          |          |          |          |       |        |           |             |
| **TryAddWithoutValidationTests** | **PR** | **Hashset**             | **73.48 ns** | **0.945 ns** | **0.884 ns** | **73.63 ns** | **72.30 ns** | **75.03 ns** |  **0.95** | **0.0227** |     **288 B** |        **1.00** |
| TryAddWithoutValidationTests | master      | Hashset             | 77.19 ns | 0.902 ns | 0.799 ns | 77.15 ns | 76.24 ns | 79.22 ns |  1.00 | 0.0227 |     288 B |        1.00 |
|                              |                                                                                                                |                     |          |          |          |          |          |          |       |        |           |             |
| **TryAddWithoutValidationTests** | **PR** | **List**                | **36.77 ns** | **0.477 ns** | **0.446 ns** | **36.84 ns** | **35.91 ns** | **37.65 ns** |  **0.48** | **0.0101** |     **128 B** |        **0.44** |
| TryAddWithoutValidationTests | master      | List                | 76.21 ns | 0.468 ns | 0.438 ns | 76.24 ns | 75.48 ns | 77.11 ns |  1.00 | 0.0228 |     288 B |        1.00 |