@pedrobsaila 2024-05-31
The new result after the last changes (it seems we degraded the slow path by rewriting it as foreach loop that calls unitary TryAddWithoutValidation) :

| Toolchain                                                                                                      | data    | Mean     | Error    | StdDev   | Median   | Min      | Max      | Ratio | RatioSD | Gen0   | Allocated | Alloc Ratio |
|--------------------------------------------------------------------------------------------------------------- |-------- |---------:|---------:|---------:|---------:|---------:|---------:|------:|--------:|-------:|----------:|------------:|
| **PR** | **Array**   | **52.23 ns** | **0.452 ns** | **0.353 ns** | **52.18 ns** | **51.69 ns** | **52.84 ns** |  **0.68** |    **0.01** | **0.0100** |     **128 B** |        **0.46** |
| master      | Array   | 76.35 ns | 1.302 ns | 1.218 ns | 76.41 ns | 74.83 ns | 79.53 ns |  1.00 |    0.02 | 0.0222 |     280 B |        1.00 |
|         |          |          |          |          |          |          |       |         |        |           |             |
| **PR** | **Hashset** | **84.33 ns** | **3.491 ns** | **3.880 ns** | **82.85 ns** | **80.04 ns** | **92.59 ns** |  **1.10** |    **0.05** | **0.0229** |     **288 B** |        **1.00** |
| master      | Hashset | 76.53 ns | 0.741 ns | 0.693 ns | 76.30 ns | 75.67 ns | 77.71 ns |  1.00 |    0.01 | 0.0229 |     288 B |        1.00 |
|                                                                                                                |         |          |          |          |          |          |          |       |         |        |           |             |
| **PR** | **List**    | **36.12 ns** | **0.289 ns** | **0.271 ns** | **36.05 ns** | **35.78 ns** | **36.68 ns** |  **0.47** |    **0.01** | **0.0102** |     **128 B** |        **0.44** |
| master      | List    | 76.49 ns | 0.829 ns | 0.692 ns | 76.42 ns | 75.49 ns | 78.12 ns |  1.00 |    0.01 | 0.0229 |     288 B |        1.00 |