@pedrobsaila 2024-06-03
New benchmark results :

| Toolchain                                                                                                      | data    | Mean     | Error    | StdDev   | Median   | Min      | Max      | Ratio | RatioSD | Gen0   | Allocated | Alloc Ratio |
|--------------------------------------------------------------------------------------------------------------- |-------- |---------:|---------:|---------:|---------:|---------:|---------:|------:|--------:|-------:|----------:|------------:|
| **PR** | **Array**   | **51.95 ns** | **0.638 ns** | **0.498 ns** | **52.16 ns** | **51.05 ns** | **52.46 ns** |  **0.68** |    **0.01** | **0.0101** |     **128 B** |        **0.46** |
| master      | Array   | 75.87 ns | 0.727 ns | 0.680 ns | 75.67 ns | 74.51 ns | 77.00 ns |  1.00 |    0.01 | 0.0223 |     280 B |        1.00 |
|                                                                                                                |         |          |          |          |          |          |          |       |         |        |           |             |
| **PR** | **Hashset** | **80.39 ns** | **0.579 ns** | **0.513 ns** | **80.30 ns** | **79.74 ns** | **81.31 ns** |  **1.03** |    **0.02** | **0.0227** |     **288 B** |        **1.00** |
| master      | Hashset | 77.68 ns | 1.145 ns | 1.071 ns | 77.89 ns | 75.62 ns | 79.58 ns |  1.00 |    0.02 | 0.0227 |     288 B |        1.00 |
|                                                                                                                |         |          |          |          |          |          |          |       |         |        |           |             |
| **PR** | **List**    | **36.47 ns** | **0.373 ns** | **0.349 ns** | **36.43 ns** | **35.98 ns** | **37.20 ns** |  **0.47** |    **0.01** | **0.0101** |     **128 B** |        **0.44** |
| master      | List    | 77.63 ns | 1.135 ns | 1.006 ns | 77.47 ns | 75.94 ns | 79.15 ns |  1.00 |    0.02 | 0.0227 |     288 B |        1.00 |